fix: Update all tests
